WEBST-2 - Repositories and controllers in data-api created.

+++ storage/dataapi/src/main/java/com/keczaps/webstaurant/users/UserController.java
@@ -1,17 +1,20 @@
-package com.keczaps.webstaurant.dataapi.users;
+package com.keczaps.webstaurant.users;
 
-import com.keczaps.webstaurant.dataapi.configuration.DataApiConfiguration;
+import com.google.common.collect.Lists;
 import com.keczaps.webstaurant.user.User;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Sort;
-import org.springframework.data.mongodb.core.MongoTemplate;
+import org.springframework.format.annotation.DateTimeFormat;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.ArrayList;
+import java.time.LocalDateTime;
+import java.time.format.DateTimeFormatter;
+import java.util.Date;
 import java.util.List;
 import java.util.UUID;
 
@@ -19,12 +22,11 @@ import java.util.UUID;
 @RequiredArgsConstructor(onConstructor = @__(@Autowired))
 public class UserController {
 
-    private final MongoTemplate mongoTemplate;
     private final UserRepository userRepository;
 
     @GetMapping(path = "/users", produces = MediaType.APPLICATION_JSON_VALUE)
     public ResponseEntity getUsers(Sort sort) {
-        List<User> users = userRepository.findAll();
+        List<User> users = userRepository.findAll(sort);
         return ResponseEntity.ok(users);
     }
 
@@ -41,6 +43,13 @@ public class UserController {
             consumes = MediaType.APPLICATION_JSON_VALUE)
     public ResponseEntity createUser(@RequestBody User user) {
         user.setId(UUID.randomUUID().toString());
+        user.setAccountNonExpired(true);
+        user.setAccountNonLocked(true);
+        user.setCredentialsNonExpired(true);
+        user.setEnabled(true);
+        user.setCreatedAt(LocalDateTime.now());
+        user.setAuthorities(Lists.newArrayList(new SimpleGrantedAuthority("ROLE_USER")));
+        user.setUserRank(User.Rank.FLAVOR_AMATOUR);
         userRepository.save(user);
         return ResponseEntity.ok(user);
     }
@@ -52,16 +61,20 @@ public class UserController {
         if (userMongoInstance == null) {
             return ResponseEntity.notFound().build();
         }
-        // TODO - user validation
+        // TODO - user validation - REST API
         userMongoInstance = user;
         userRepository.save(userMongoInstance);
-        return new ResponseEntity(user, HttpStatus.CREATED);
+        return ResponseEntity.status(HttpStatus.CREATED).body(user);
     }
 
-    @RequestMapping(path = "/users/{id}", method = RequestMethod.DELETE,
-            produces = MediaType.APPLICATION_JSON_VALUE, consumes = MediaType.APPLICATION_JSON_VALUE)
+    @DeleteMapping(path = "/users/{id}", produces = MediaType.APPLICATION_JSON_VALUE, consumes = MediaType.APPLICATION_JSON_VALUE)
     public ResponseEntity deleteUser(@PathVariable("id") String id) {
+        User user = userRepository.findOne(id);
+        if (user == null){
+            return ResponseEntity.notFound().build();
+        }
         userRepository.delete(id);
-        return new ResponseEntity(HttpStatus.OK);
+        return ResponseEntity.ok(user);
     }
+
 }
